Refactor: 공통 컴포넌트 Directory 변경
